fix(plugin-core): keep asset path helper out of core
Move deterministic storage key derivation into plugin-core so the shared React Native core package does not grow server/plugin storage API surface.

Constraint: @hot-updater/core is consumed by the React Native runtime package.

Rejected: Leaving the helper in core | it is not Node-only, but it exposes storage-provider concerns from the runtime-facing package.

Confidence: high

Scope-risk: narrow

Directive: Keep provider/server storage helpers in plugin-core unless they are required by RN runtime code.

Tested: pnpm -w exec vitest run packages/hot-updater/src/commands/deploy.spec.ts packages/server/src/db/createBundleDiff.spec.ts packages/server/src/db/pluginCore.spec.ts plugins/cloudflare/src/r2Storage.spec.ts

Tested: pnpm --filter @hot-updater/plugin-core test:type && pnpm --filter @hot-updater/server test:type && pnpm --filter hot-updater test:type

Not-tested: Remote provider bucket race under concurrent deploys
